work on member list
