+++ src/app/member/components/MemberActionRender.tsx
@@ -1,4 +1,4 @@
-import ActionsMenu from '@/components/common/ActionMenu/ActionsMenu';
+import ActionsMenu from '@/common/ActionMenu/ActionsMenu';
 import { MemberRole } from '@/lib/routes';
 import { Info, Trash2, Edit, Send } from "lucide-react";
 import React from 'react';
